requirements.txt
PyYAML==6.0
email-validator==1.2.1
fastapi==0.75.2
gunicorn==20.1.0
pandas==1.4.2
passlib==1.7.4
psutil==5.9.0
psycopg2==2.9.5
pyarrow==11.0.0
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
retry==0.9.2
sqlalchemy==1.4.36
sqlalchemy_utils==0.38.2
uvicorn==0.17.6
requests-mock==1.10.0
polars==0.18.3
